* TimeoutProvider: Warning in MSVC removed  * Thread: Implemented existing join method for MSVC and added new join variant; Added sleep method ; Added some documentation.  * Timestamp: Fixed compile problem for MSVC  * Mutex: Fixed bug in destructor if compiling with MSVC ; clean up  * Added examples directory with the purpose of testing features individually (simplifies porting)  * Added example/test of the Mutex class  * Added example/test of the Semaphore class  * Added example/test of the Thread class  * MSVC Makefile: Changed to dynamic linking  * config.h: Removed one warning "W4251". We need to find out how to avoid it. This helps seeing    other warnings. Remove this fix later.
git-svn-id: svn://svn.minisip.org/minisip/trunk@1334 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libmutil/include/libmutil/TimeoutProvider.h
@@ -110,7 +110,7 @@ class TPRequest{
 			if (happens_before(now))
 				return 0;
 			else
-				return when_ms - now;
+				return (int)(when_ms - now);
 		}
 		
 		TOCommand get_command(){
